# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AgroConsult API documentation",
        "description": "AgroConsult API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/resume": {
            "get": {
                "tags": [
                    "Resume"
                ],
                "summary": "Resume list",
                "description": "index",
                "operationId": "ResumeIndex",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Resume"
                ],
                "summary": "Create new resume",
                "description": "Create ",
                "operationId": "48dc484527e77d29399421935ac806c1",
                "requestBody": {
                    "description": "Create new Resume",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "sub_category_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "language": {
                                        "type": "text",
                                        "example": "['english', 'russian']"
                                    },
                                    "about": {
                                        "type": "string",
                                        "example": "AgroConsult"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, filled input. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/resume/{resume}": {
            "get": {
                "tags": [
                    "Resume"
                ],
                "description": "Show by Resume",
                "operationId": "ResumeShow",
                "parameters": [
                    {
                        "name": "resume",
                        "in": "path",
                        "description": "Resume ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Resume"
                ],
                "summary": "Update Resume",
                "description": "Update ",
                "operationId": "22629d84ce845a3ed32b247254338992",
                "parameters": [
                    {
                        "name": "resume",
                        "in": "path",
                        "description": "Resume ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Category",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "profile_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "category_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "sub_category_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "language": {
                                        "type": "text",
                                        "example": "['english', 'russian']"
                                    },
                                    "about": {
                                        "type": "string",
                                        "example": "AgroConsult"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Resume"
                ],
                "summary": "delete Resume",
                "description": "delete ",
                "operationId": "803ac1e08ba3f50437502d0682fbc2aa",
                "parameters": [
                    {
                        "name": "resume",
                        "in": "path",
                        "description": "Resume ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}